# Conflicts:
#	src/main/java/fr/epsi/mspr/entities/Product.java
#	src/main/java/fr/epsi/mspr/entities/Purchase.java
#	src/main/java/fr/epsi/mspr/repositories/ProductInPurchaseRepository.java
#	src/main/java/fr/epsi/mspr/repositories/ProductRepository.java
#	src/main/resources/templates/commandes.html

+++ src/main/java/fr/epsi/mspr/repositories/ProductRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
 
-    List<Product> findByCategoryId(int category_id);
+    List<Product> findByCategoryId(Long category_id);
 
     //Update the quantity of product
     @Transactional
